requirements.txt
streamlit
pillow
opencv-python
ultralytics
torch
torch-vision
numpy==1.26.4